Moved import/export code to new tool.
Fixed external storage access bug for Android 10.
Added capability to export to eml.
