+++ Mifare Classic Tool/app/src/main/AndroidManifest.xml
@@ -32,6 +32,7 @@
     <application
         android:name="de.syss.MifareClassicTool.Common"
         android:allowBackup="false"
+        android:requestLegacyExternalStorage="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
